requirements.txt
streamlit
google.generativeai
langchain
langchain.llms
google.cloud